refactor: switch plugin siblings to self-reexport imports
