+++ packages/opencode/src/cli/cmd/tui/plugin/runtime.ts
@@ -24,8 +24,8 @@ import {
   type PluginPackage,
   type PluginSource,
 } from "@/plugin/shared"
-import { PluginLoader } from "@/plugin"
-import { PluginMeta } from "@/plugin"
+import { PluginLoader } from "@/plugin/loader"
+import { PluginMeta } from "@/plugin/meta"
 import { installPlugin as installModulePlugin, patchPluginConfig, readPluginManifest } from "@/plugin/install"
 import { hasTheme, upsertTheme } from "../context/theme"
 import { Global } from "@/global"
