--- packages/opencode/src/plugin/github-copilot/index.ts
@@ -1,1 +1,0 @@
-export * as CopilotModels from "./models"
