+++ packages/opencode/src/plugin/github-copilot/models.ts
@@ -142,3 +142,4 @@ export async function get(
 
   return result
 }
+export * as CopilotModels from "./models"
