+++ packages/opencode/src/plugin/index.ts
@@ -1,3 +1,1 @@
 export * as Plugin from "./plugin"
-export * as PluginMeta from "./meta"
-export * as PluginLoader from "./loader"
